+++ scripts/deploy-in-vscode.js
@@ -3,7 +3,7 @@
 // npm-groovy-lint & vscode-groovy-lint must be in the same folder, else you have to override  VSCODE_GROOVY_LINT_PATH env variable
 
 // Imports
-import fs from "fs-extra";
+import fs from "node:fs";
 
 console.log("---- START DEPLOY IN VSCODE INSTALLED npm-groovy-lint PACKAGE ----");
 
@@ -14,15 +14,12 @@ const targetPath = `${vsCodeGroovyLintPath}/server/node_modules/npm-groovy-lint`
 console.info(`GroovyLint: Starting copying package in vscode for testing`);
 
 // Reset target folder
-if (fs.existsSync(targetPath)) {
-    fs.emptyDirSync(targetPath);
-} else {
-    fs.mkdirSync(targetPath);
-}
+fs.rmSync(targetPath, { recursive: true, force: true });
+fs.mkdirSync(targetPath, { recursive: true });
 
 // Copy files into dest folder
 for (const path of ["package.json", "README.md", "CHANGELOG.md", "LICENSE", "index.d.ts", "lib"]) {
-    fs.copySync(path, `${targetPath}/${path}`);
+    fs.cpSync(path, `${targetPath}/${path}`, { recursive: true });
 }
 
 console.info(`GroovyLint: Copied files into ${targetPath}`);