+++ test/errors.test.js
@@ -1,7 +1,7 @@
 #! /usr/bin/env node
 import NpmGroovyLint from "../lib/groovy-lint.js";
 import assert from "assert";
-import fs from "fs-extra";
+import fs from "node:fs";
 import { SAMPLE_FILE_SMALL_PATH } from "./helpers/common.js";
 
 describe("Errors", function () {