+++ test/lint-fix.test.js
@@ -1,8 +1,7 @@
 #! /usr/bin/env node
 import NpmGroovyLint from "../lib/groovy-lint.js";
 import assert from "assert";
-import fs from "fs-extra";
-import * as rimraf from "rimraf";
+import fs from "node:fs";
 import { beforeEachTestCase, copyFilesInTmpDir, checkCodeNarcCallsCounter, SAMPLE_FILE_BIG_PATH, SAMPLE_FILE_SMALL_PATH } from "./helpers/common.js";
 
 describe("Lint & fix with API", function () {
@@ -138,8 +137,8 @@ describe("Lint & fix with API", function () {
             assert(!linter.outputString.includes("NaN"), "Results does not contain NaN");
             checkCodeNarcCallsCounter(2);
         } finally {
-            fs.removeSync("npm-groovy-fix-log.json");
-            rimraf.sync(tmpDir);
+            fs.rmSync("npm-groovy-fix-log.json", { recursive: true, force: true });
+            fs.rmSync(tmpDir, { recursive: true, force: true });
         }
     }).timeout(120000);
 
@@ -191,8 +190,8 @@ describe("Lint & fix with API", function () {
             assert(!linter.outputString.includes("NaN"), "Results does not contain NaN");
             checkCodeNarcCallsCounter(1);
         } finally {
-            fs.removeSync("npm-groovy-fix-log-should-fix-only-some-errors.txt");
-            rimraf.sync(tmpDir);
+            fs.rmSync("npm-groovy-fix-log-should-fix-only-some-errors.txt", { recursive: true, force: true });
+            fs.rmSync(tmpDir, { recursive: true, force: true });
         }
     }).timeout(120000);
 
@@ -234,8 +233,8 @@ describe("Lint & fix with API", function () {
 
             checkCodeNarcCallsCounter(1);
         } finally {
-            fs.removeSync("npm-groovy-fix-log-should-exclude-some-rules.txt");
-            rimraf.sync(tmpDir);
+            fs.rmSync("npm-groovy-fix-log-should-exclude-some-rules.txt", { recursive: true, force: true });
+            fs.rmSync(tmpDir, { recursive: true, force: true });
         }
     }).timeout(120000);
 
@@ -264,8 +263,8 @@ describe("Lint & fix with API", function () {
             assert(!linter.outputString.includes("NaN"), "Results does not contain NaN");
             checkCodeNarcCallsCounter(2);
         } finally {
-            fs.removeSync("npm-groovy-fix-log-should-fix-groovy-files.txt");
-            rimraf.sync(tmpDir);
+            fs.rmSync("npm-groovy-fix-log-should-fix-groovy-files.txt", { recursive: true, force: true });
+            fs.rmSync(tmpDir, { recursive: true, force: true });
         }
     }).timeout(120000);
 });